\name{aermod-package}
\alias{aermod-package}
\alias{aermod-package}
\title{Reads data from formats used by AERMOD}
\description{Interface to the AERMOD dispersion model}
\details{\tabular{ll}{
Package: \tab aermod\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-06-03\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}}
\alias{aermod}
\docType{package}
\author{David Holstius}
\references{\url{http://www.epa.gov/ttn/scram/dispersion_prefrec.htm#aermod}}
\keyword{package}
\seealso{\code{\link{read.unformatted}}}

