% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend}
\alias{blend}
\title{Blend two impact matrices together.}
\usage{
blend(src, indices, target)
}
\arguments{
\item{src}{Matrix corresponding to source group for which intermittent operation is simulated.}

\item{indices}{Hours of operation (integer vector). Typically generated with \code{hrs_rand}}

\item{target}{Matrix representing source group that does \emph{not} operate when intermittent source does. By default this is the zero matrix (i.e., intermittent source may operate at same time as all other sources).}
}
\value{
Blended matrix. Has an attribute listing all of the hours for which a substitution was made. Existing attributes are carried over from \code{target}.
}
\description{
The \code{blend} operation inserts rows from a source impact matrix (typically corresponding to an impact matrix for a non-continuous source)
into a target impact matrix (typically corresponding to a continuous source that does not operate at the same time as the non-continuous source).
Blend also generates an impact matrix which is a zero matrix of the same dimensions as the source matrix, except for the selected rows.
}
\details{
Indices are intended to be a set of randomly generated operating hours created by \code{hrs_rand} (which may be created via \code{make_simplan}).
}
