% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_tbl.R
\name{control_tbl}
\alias{control_tbl}
\title{Control options}
\usage{
control_tbl(title = "Demo Energy Station",
  subtitle = "Default Run: Hour, Day, Annual", model_opt = c("DFAULT",
  "CONC"), avg_time = c("1", "24", "ANNUAL"), urban_opt = as.character(NA),
  pollutant_id = "OTHER", flagpole = as.character(NA),
  event_file = as.character(NA), error_file = as.character(NA),
  multi_year = as.character(NA), debug_opt = as.character(NA))
}
\arguments{
\item{title}{Title of modeling run. `TITLEONE` of AERMOD input file.}

\item{subtitle}{Subtitle of modeling run. `TITLETWO` of AERMOD input file.}

\item{model_opt}{Model run options. 
Enter multiple options as a vector: \code{c("DFAULT", "CONC")}. 
For a complete list of options, see 'MODELOPT' in the AERMOD reference guide at 
https://www3.epa.gov/ttn/scram/models/aermod/aermod_quick-reference-guide.pdf.}

\item{avg_time}{Dispersion factor averaging times. Enter multiple options as a vector: \code{c("1", "24", "ANNUAL")}.}

\item{urban_opt}{Urban options for AERMOD run. Enter multiple options as a vector: \code{c("400000", "Minneapolis")}.}

\item{pollutant_id}{Pollutant ID. Options: (1) "NO2"  (2) "SO2"  (3) "OTHER"...}

\item{flagpole}{Flagpole options.}

\item{event_file}{File to write AERMOD 'events' to.}

\item{error_file}{File to write AERMOD errors to.}

\item{multi_year}{Options for performing multiple "year" AERMOD runs. 
Use for identifying the highest annual value across multiple years of met data.}

\item{debug_opt}{Automatic debugging options performed by AERMOD.}
}
\description{
Create a data frame of AERMOD control options.
}
\examples{
control_tbl(title = "My New Space Station")

}
\keyword{aermod}
\keyword{control}
\keyword{input}
\keyword{options}

