% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downwash_df.R
\name{downwash_df}
\alias{downwash_df}
\title{Downwash options}
\usage{
downwash_df(source_id = as.character(NA), bld_heights = as.numeric(NA),
  bld_widths = as.numeric(NA), bld_lengths = as.numeric(NA),
  x_badj = as.numeric(NA), y_badj = as.numeric(NA))
}
\arguments{
\item{source_id}{Emission source ID or name.}
}
\description{
Create a data frame of AERMOD downwash options.
}
\examples{
downwash_df(source_id = "STK_1")
}
\keyword{aermod}
\keyword{building}
\keyword{downwash}
\keyword{input}

