% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downwash_tbl.R
\name{downwash_tbl}
\alias{downwash_tbl}
\title{Downwash options}
\usage{
downwash_tbl(source_id = NULL, build_heights = NULL, build_widths = NULL,
  build_lengths = NULL, x_badj = NULL, y_badj = NULL)
}
\arguments{
\item{source_id}{Emission source ID or name.}
}
\description{
Create a data frame of AERMOD downwash options.
}
\examples{
downwash_tbl(source_id = "STK_1")
}
\keyword{aermod}
\keyword{building}
\keyword{downwash}
\keyword{input}

