% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_so2_dv.R
\name{fast_so2_dv}
\alias{fast_so2_dv}
\title{Calculate 2010 SO2 design values using base functions only}
\usage{
fast_so2_dv(mat, sumfn = identity)
}
\arguments{
\item{mat}{An impact matrix over five years of met data}

\item{sumfn}{Use \code{max} for max DV at any receptor, default is \code{identity} (DV at each receptor, in order).}
}
\value{
A vector of design values.
}
\description{
Calculate 2010 SO2 design values using base functions only
}
