% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max.R
\name{get_max}
\alias{get_max}
\title{Get max dispersion result from AERMOD output file}
\usage{
get_max(data = as.character(NA), property_boundary = as.numeric(NA))
}
\description{
Read maximum dispersion results from AERMOD output file into a data frame.
}

