\name{grid.polar}
\alias{grid.polar}
\title{grid.polar}
\usage{grid.polar(r, n, offset, increment, units="degrees", origin=c(0, 0))}
\description{Generate a polar grid.}
\value{a matrix}
\references{EPA (2004). User's Guide for the AMS/EPA Regulatory Model AERMOD. EPA-454/B-03-001.}
\arguments{\item{r}{distances (arbitrary units)}
\item{n}{angles}
\item{offset}{starting angle}
\item{increment}{angualr increment}
\item{units}{(optional) angular units ("degrees" or "radians")}
\item{origin}{(optional) x, y offset}}
