% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_over.R
\name{is_over}
\alias{is_over}
\title{Check whether a point falls over a given land use type}
\usage{
is_over(lat, long, category)
}
\arguments{
\item{lat}{Latitude as WGS84 (e.g. 46.34)}

\item{long}{Longitude as WGS84 (e.g. -90.34)}

\item{category}{Land use category to check for.  Can be "developed", "undeveloped", "agriculture", "forest", "shrub", or "water".}
}
\description{
Input a lat, long location and a land use category and return
TRUE if the point falls within that category
}
\examples{
is_over(lat = 45.121, long = -96.342, category = "developed")
is_over(lat = 46.152, long = -93.720, category = "undeveloped")
is_over(lat = 46.152, long = -93.720, category = "water")
is_over(lat = 46.139, long = -93.705, category = "water")
is_over(lat = 46.071, long = -93.668, category = "developed")
is_over(lat = 46.071, long = -93.668, category = "undeveloped")
}
\keyword{is_over}
\keyword{land}
\keyword{landuse}
\keyword{use}
