% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Test if object is valid}
\usage{
is_valid(x, n = 1)
}
\arguments{
\item{x}{Object for length test.}

\item{n}{Minimum number of characters if character or numeric; 
Minimum number of rows if data frame.}
}
\description{
Test for nchars() or nrows() of object. 
For lists, returns \code{TRUE} if
at least one member of the list passes validity test.
\code{NA} objects return \code{FALSE}.
}
\examples{
# For a character string
is_valid("MyName", n = 3)

is_valid(as.character(NA), n = 1)

# For a data frame
is_valid(data.frame(X = 1), n = 1)

is_valid(data.frame(X = 1), n = 2)

# For numeric
is_valid(13, n = 2)

# For a list
is_valid(list(NA, "MyName"), n = 3)

}
\keyword{minimum}
\keyword{test}
\keyword{valid}
