% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_to_utm.R
\name{ll_to_utm}
\alias{ll_to_utm}
\title{Convert Geodetic Coordinates (Longitude, Latitude) to UTM Coordinates}
\usage{
ll_to_utm(df, long = "LONG", lat = "LAT", utm_zone = 15)
}
\arguments{
\item{df}{Input data frame.}

\item{long}{The name of the column containing longitude values}

\item{lat}{The name of the column containing latitude values}

\item{utm_zone}{UTM zone in which to calculate X, Y coordinates}
}
\description{
Convert longitude, latitude coordinates in a data frame to UTM coordinates
in a specified UTM zone.  The longitude and latitude coordinates should be
columns in the data frame passed into the function.  The user can optionally
specify the names of the columns to process.  The function returns the
dataframe with the X and Y columns bound to it.
}
\examples{
ll_to_utm(df, long = "LONG", lat = "LAT", utm_zone = 15)
}
\keyword{ll}
\keyword{utm}
