% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simplan.R
\name{make_simplan}
\alias{make_simplan}
\title{Generate a list of random operating hours, along with
the PRNG seeds needed to recreate them.}
\usage{
make_simplan(N_sim = 5, ...)
}
\arguments{
\item{N_sim}{The number of sets of operating hours to generate.}

\item{...}{Input parameters for call to \code{hrs_rand}.}
}
\value{
a nested list of N_sim elements, each of which is a list with elements \code{seed} and \code{hrs}
}
\description{
Generate a list of random operating hours, along with
the PRNG seeds needed to recreate them.
}
