% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_df.R
\name{met_df}
\alias{met_df}
\title{Meteorology options}
\usage{
met_df(
  surf_file = as.character(NA),
  surf_site_info = as.character(NA),
  prof_file = as.character(NA),
  upper_air_info = as.character(NA),
  base_elev_m = as.numeric(NA),
  start_date = as.character(NA),
  end_date = as.character(NA)
)
}
\arguments{
\item{surf_file}{Surface met data file location.}

\item{surf_site_info}{Optional surface met site description: \code{"Station_no Year City/Name"}.}

\item{prof_file}{Profile met data file location.}

\item{upper_air_info}{Optional upper air site description: \code{"Station_no Year City/Name"}.}

\item{base_elev_m}{Base elevation in meters above mean sea level of temperature profile data.}

\item{start_date}{Start date of meteorology data to be modeled. 
Enter as text: \code{"year month day hour"}
If blank, set to the first entry of \code{surf_file}.}

\item{end_date}{End date of meteorology data to be modeled. 
Enter as text : \code{"year month day hour"}
If blank, set to the last entry of \code{surf_file}.}
}
\description{
Create a data frame of AERMOD meteorology options.
}
\examples{
met_df(surf_file      = "..\\\\MET\\\\RSTMPX5Y.SFC",
        surf_site_info = "14925 2006 ROCHESTER/MUNICIPAL_ARPT",
        prof_file      = "..\\\\MET\\\\RSTMPX5Y.PFL",
        upper_air_info = "94983 2006 ROCHESTER/MUNICIPAL_ARPT",
        base_elev_m    = 396.0,
        start_date     = "2009 1 1 1",
        end_date       = "2013 12 31 24")
}
\keyword{aermod}
\keyword{input}
\keyword{meteorology}
\keyword{surface}
\keyword{weather}
