% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_tbl.R
\name{out_tbl}
\alias{out_tbl}
\title{Output options}
\usage{
out_tbl(rect_table = "ALLAVE 1ST, 1 1ST", max_table = as.character(NA),
  day_table = as.character(NA), plot_file = as.character(NA))
}
\arguments{
\item{rect_table}{Output options as text. Separate multiple outputs with a comma: \code{"ALLAVE 1ST, 1 1ST"}}

\item{max_table}{Output options as text.}

\item{day_table}{Output options as text.}

\item{plot_file}{Output options as text.}
}
\description{
Create an input table of AERMOD output options.
}
\examples{
out_tbl(rect_table = "ALLAVE 1ST, 1 1ST 8TH")
}
\keyword{aermod}
\keyword{output}

