% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_df.R
\name{proj_df}
\alias{proj_df}
\title{Projection details}
\usage{
proj_df(projection = "Coordinate System UTM",
  description = "UTM: Universal Transverse Mercator",
  datum = "North American datum 1983", dtmrgn = "conus", units = "m",
  zone = "15", zone_inx = "0")
}
\arguments{
\item{projection}{Name of coordinate projection used for receptors.}

\item{description}{Extended description of coordinate projection used for receptors.}

\item{zone}{UTM zone of coordinates.}
}
\description{
Create a data frame of projection details for AERMOD receptors.
}
\examples{
proj_df(projection  = "Coordinate System UTM")
}
\keyword{aermod}
\keyword{options}
\keyword{output}
\keyword{projection}
