% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_tbl.R
\name{proj_tbl}
\alias{proj_tbl}
\title{Projection details}
\usage{
proj_tbl(projection = "Coordinate System UTM",
  description = "UTM: Universal Transverse Mercator",
  datum = "North American datum 1983", dtmrgn = "conus", units = "m",
  zone = "15", zoneinx = "0")
}
\arguments{
\item{projection}{Coordinate projection of receptors.}

\item{description}{Coordinate projection.}

\item{zone}{UTM zone of coordinates.}
}
\description{
Create a data frame of projection details for AERMOD receptors.
}
\examples{
proj_tbl(projection  = "Coordinate System UTM")
}
\keyword{aermod}
\keyword{input}
\keyword{options}
\keyword{projection}

