% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aermod_out.R
\name{read_aermod_out}
\alias{read_aermod_out}
\title{Read AERMOD output file}
\usage{
read_aermod_out(file)
}
\arguments{
\item{file}{File location. Defaults to "aermod.out".}
}
\description{
Read an aermod.out file into a dataframe.
}
\examples{
\dontrun{
read_aermod_out(file = "aermod.out")
}
}
\keyword{aermod}
\keyword{output}
\keyword{read}
\keyword{results}
