% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plt.R
\name{read_plt}
\alias{read_plt}
\title{Read AERMOD PLOT Output Files}
\usage{
read_plt(infile)
}
\arguments{
\item{infile}{Input PLT file to read.}
}
\description{
Read a table of AERMOD receptor concentration data from the
PLOTFILE option into a \code{\link{data.frame}}.
}
\examples{
read_plt("project/pm25_srcgrp1.PLT")
}
\keyword{plot}
\keyword{plt}
