% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pos.R
\name{read_pos}
\alias{read_pos}
\title{Read AERMOD POST Output Files}
\usage{
read_pos(infile)
}
\arguments{
\item{infile}{Input pos file to read.}
}
\description{
Read a table of AERMOD post processed concentration data from the
POSTFILE option into a \code{\link{data.frame}}.
}
\examples{
read_pos("project/pm25_srcgrp1.POS")
}
\keyword{pos}
\keyword{post}
