% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_df.R
\name{receptor_df}
\alias{receptor_df}
\title{Receptor options}
\usage{
receptor_df(recept_file = as.character(NA),
  recept_as_text = as.character(NA))
}
\arguments{
\item{recept_file}{Receptor file location.}

\item{recept_as_text}{Additional receptor locations as text string.}
}
\description{
Create a data frame of AERMOD receptor options.
}
\examples{
\dontrun{
receptor_df(recept_file = "..\\\\Receptors\\\\circleReceptors.ROU")
}
}
\keyword{aermod}
\keyword{input}
\keyword{receptors}
