% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_tbl.R
\name{receptor_tbl}
\alias{receptor_tbl}
\title{Receptor options}
\usage{
receptor_tbl(rect_file = "", rou_text = "")
}
\arguments{
\item{rect_file}{Receptor file location.}

\item{rou_text}{Additional receptor locations as text string. Default is blank.}
}
\description{
Create an input table of AERMOD receptor options.
}
\examples{
\dontrun{
receptor_tbl(rect_file = "..\\\\Receptors\\\\circleReceptors.ROU")
}
}
\keyword{aermod}
\keyword{input}
\keyword{receptors}

