% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aermod.R
\name{run_aermod}
\alias{run_aermod}
\title{Run AERMOD}
\usage{
run_aermod(input = "aermod.inp", output = input$title,
  exe_folder = "aermod")
}
\arguments{
\item{input}{Path to input file.}

\item{output}{Filename for model results. Defaults to `title` of input file.}

\item{exe_folder}{Folder containing aermod.exe.}
}
\description{
Call aermod.exe and run input file
}
\examples{
\dontrun{
run_aermod(data       = "aermod.inp", 
           output     = "aermod_results", 
           exe_folder = "aermod_exe")
}
}
\keyword{EPA}
\keyword{aermod}
\keyword{dispersion}
\keyword{model}

