% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_postfile.R
\name{scan_postfile}
\alias{scan_postfile}
\title{Scan POST file}
\usage{
scan_postfile(f)
}
\arguments{
\item{f}{character}
}
\value{
matrix
}
\description{
Reads Fortran unformatted I/O POSTFILE and conver to matrix.
Hours are not written individually, nor are source groups
as they are saved as attributes of the matrix.
}
