% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_post.R
\name{serialize_post}
\alias{serialize_post}
\title{Serialize impact matrix}
\usage{
serialize_post(mat, fname, srcgrp)
}
\arguments{
\item{mat}{matrix}

\item{fname}{character}

\item{srcgrp}{character}
}
\value{
logical
}
\description{
Serialize impact matrix to AERMOD unformatted binary POST file.
}
