% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sher_lambda.R
\name{sher_lambda}
\alias{sher_lambda}
\title{Report Lambda Values for State Health Effects Review}
\usage{
sher_lambda(mat, esl)
}
\arguments{
\item{mat}{An impact matrix over one year of met data}

\item{an}{esl against which to report lambda values}
}
\value{
A list of various data tables useful for running Monte Carlo simulations, or reporting unadjusted lambda values.
}
\description{
Generates various statistics used in TCEQ Toxicology Tier III RFC reviews.
If the impact matrix covers more than one year of met data, only the first year is used.
}
