% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sher_monte_carlo.R
\name{sher_monte_carlo}
\alias{sher_monte_carlo}
\title{Generate 99.9 percentile value of total exceedances (10,000 Monte Carlo trials)
for State Health Effects review, given table of exceedances.}
\usage{
sher_monte_carlo(lambda, nhr, nsim = 10000L)
}
\arguments{
\item{lambda}{A lambda object generated by the \code{sher_lambda} function}

\item{nhr}{A proposed limit on hr/yr of operatoin.}
}
\value{
A table of 99.9\% percentile exceedance figures for different window sizes and ESL thresholds (2xESL, 4xESL)
}
\description{
Generates various statistics used in TCEQ Toxicology Tier III RFC reviews.
If the impact matrix covers more than one year of met data, only the first year is used.
}
