% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_df.R
\name{source_df}
\alias{source_df}
\title{Source options}
\usage{
source_df(
  source_id = as.character(NA),
  description = as.character(NA),
  elevation_m = as.numeric(NA),
  x_coord = as.numeric(NA),
  y_coord = as.numeric(NA),
  type = c("point"),
  emit_gs = as.numeric(NA),
  height_m = as.numeric(NA),
  temp_k = as.numeric(NA),
  velocity_ms = as.numeric(NA),
  diameter_m = as.numeric(NA),
  downwash_file = as.character(NA),
  urban_pop = as.numeric(NA),
  group_id = as.character(NA)
)
}
\arguments{
\item{source_id}{Emission source ID or name.}

\item{description}{Short description of emission source as a text string.}

\item{elevation_m}{Source elevation in meters.}

\item{x_coord}{X coordinates of source.}

\item{y_coord}{Y coordinates of source.}

\item{type}{Type of source to be modeled in AERMOD. 
Options: (1) "point", (2) "volume", or (3) "area".}

\item{emit_gs}{Emission rate for source in \code{grams/second}.}

\item{height_m}{Height of source in meters.}

\item{temp_k}{Exit temperature of air emissions in degrees \code{Kelvin}.}

\item{velocity_ms}{Exit velocity of air emissions in in \code{m/s}.}

\item{diameter_m}{Internal diameter of stack in \code{meters}.}

\item{downwash_file}{Path to downwash parameters for each emission source.}

\item{urban_pop}{Urban population for source. Requires a value of at least \code{100}.
If blank or a value less than \code{100}, source is modeled with default \code{rural} dispersion coefficients.}

\item{group_id}{Group IDs assigned to source. 
To assign a source to multiple groups use a vector: \code{c("ALL", "SV01", "SV01_to_SV03"}.
If blank, source is assigned to group "All".}
}
\description{
Create a data frame of AERMOD source options.
}
\examples{
source_df(source_id     = c("SV01"),
          description   = c("boiler1"),
          elevation_m   = c(361.33),
          x_coord       = c(461874.9),
          y_coord       = c(4631367),
          type          = c("point"),
          emit_gs       = c(1),
          height_m      = c(10),
          temp_k        = c(600),
          velocity_ms   = c(30),
          diameter_m    = c(2),
          downwash_file = as.character(NA),
          urban_pop     = 60000,
          group_id      = as.character(NA))
}
\keyword{aermod}
\keyword{input}
\keyword{source}
