% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Recombine receptor-disjoint POST files}
\usage{
stitch(srcgrp_idx, path = ".", ...)
}
\arguments{
\item{srcgrp_idx}{Source-group specific numeric index. I.e., \code{srcgrcp_idx=12} will pick out PST012_xxxx, which are the files for the 12th defined source group}
}
\value{
Combined POST file
}
\description{
Parallel AERMOD runs using N workers generate an ordered set of POSTFILEs
that need to be recombined. The POSTFILEs all correspond to the same met
database and therefore have the same number of rows. The receptors (columns)
need to be re-bound in order.
}
\details{
The POST are all named in a standard way such that PST003_0012 is the twelfth
disjoint POST file corresponding to the thid source group.
}
