% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_to_ll.R
\name{utm_to_ll}
\alias{utm_to_ll}
\title{Convert UTM Coordinates to Geodetic Coordinates (Longitude, Latitude)}
\usage{
utm_to_ll(df, utm_zone = 15, x = "X", y = "Y")
}
\arguments{
\item{df}{Input data frame.}

\item{utm_zone}{UTM zone in which to calculate X, Y coordinates}

\item{long}{The name of the column containing longitude values}

\item{lat}{The name of the column containing latitude values}
}
\description{
Convert UTM coordinates in a specified UTM zone in a data frame to 
longitude, latitude coordinates. The UTM coordinates should be
columns in the data frame passed into the function.  The user can optionally
specify the names of the columns to process.  The function returns the
dataframe with the LONG and LAT columns bound to it.
}
\examples{
ll_to_utm(df, long = "LONG", lat = "LAT", utm_zone = 15)
}
\keyword{ll}
\keyword{utm}
