% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aermod.R
\name{write_aermod}
\alias{write_aermod}
\title{Write a data frame to an AERMOD input file}
\usage{
write_aermod(data = NULL, path = NULL, control = NULL, sources = NULL,
  receptors = NULL, met = NULL, out = NULL)
}
\arguments{
\item{data}{Data frame of AERMOD modeling parameters.}

\item{path}{Path to write to. Default is "aermod.inp".}

\item{control}{Data frame of control parameters. If blank, control parameters from \code{data} are used.}

\item{sources}{Data frame of source parameters. If blank, source parameters from \code{data} are used.}

\item{receptors}{Data frame of receptor parameters. If blank, receptor parameters from \code{data} are used.}

\item{met}{Data frame of meteorology parameters. If blank, meteorology parameters from \code{data} are used.}

\item{out}{Data frame of output parameters. If blank, output parameters from \code{data} are used.}
}
\description{
Output an AERMOD input file from a data frame of AERMOD parameters.
}
\examples{
aermod_inp <- new_aermod()

write_aermod(data = aermod_inp, path = "aermod.inp")
}
\keyword{aermod}
\keyword{input}
\keyword{save}
\keyword{write}

