% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_tbl.R
\name{control_tbl}
\alias{control_tbl}
\title{Control options}
\usage{
control_tbl(title_one = "Demo Energy Station",
  title_two = "Default Run: Hour, Day, Annual", model_opt = "DFAULT, CONC",
  avg_time = "1, 24, ANNUAL", urban_opt = "", pollutant_id = "OTHER",
  flagpole = "", event_file = "", error_file = "", multi_year = "",
  debug_opt = "")
}
\arguments{
\item{title_one}{Title.}

\item{title_two}{Subtitle.}

\item{model_opt}{Model run options. 
Separate multiple options with a comma: \code{"DFAULT, CONC"}. 
For a complete list of options, see 'MODELOPT' in the AERMOD reference guide at 
https://www3.epa.gov/ttn/scram/models/aermod/aermod_quick-reference-guide.pdf.}

\item{avg_time}{Dispersion factor averaging times. Separate multiple options with a comma: \code{"1, 24, ANNUAL"}.}

\item{urban_opt}{Urban options for AERMOD run. Separate multiple options with a comma: \code{"500000 Minneapolis"}.}

\item{pollutant_id}{Pollutant ID. Options: (1) "NO2" (2)...}

\item{flagpole}{Flagpole options.}

\item{event_file}{File to write AERMOD 'events' to.}

\item{error_file}{File to write AERMOD errors to.}

\item{multi_year}{Optoons for performing multiple "year" AERMOD runs. 
Used for identifying the highest annual value across multiple years of met data.}

\item{debug_opt}{Automatic debugging options performed by AERMOD.}
}
\description{
Create an input table of AERMOD control options.
}
\examples{
control_tbl(title_one = "Demo Energy Station")

}
\keyword{aermod}
\keyword{control}
\keyword{input}
\keyword{options}

