% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_tbl.R
\name{control_tbl}
\alias{control_tbl}
\title{Control options}
\usage{
control_tbl(TITLEONE = "Demo Energy Station", TITLETWO = "AERMOD Run",
  MODELOPT = "DFAULT, CONC", AVERTIME = "1, 24, ANNUAL", URBANOPT = "",
  POLLUTID = "OTHER", HALFLIFE = "", DCAYCOEF = "", FLAGPOLE = "",
  EVENTFIL = "", SAVEFILE = "", INITFILE = "", MULTYEAR = "",
  DEBUGOPT = "", ERRORFIL = "", RUNORNOT = TRUE)
}
\arguments{
\item{TITLEONE}{Title of AERMOD run.}

\item{MODELOPT}{AERMOD run options. Separate multiple options by comma.}
}
\description{
Create an input table of AERMOD control options.
}
\examples{
control_tbl(TITLEONE = "Demo Energy Station")

TITLEONE            TITLETWO   MODELOPT     AVERTIME      URBANOPT POLLUTID HALFLIFE DCAYCOEF FLAGPOLE EVENTFIL SAVEFILE INITFILE MULTYEAR DEBUGOPT ERRORFIL RUNORNOT
Demo Energy Station AERMOD Run DFAULT, CONC 1, 24, ANNUAL          OTHER                                                                                     TRUE 
}
\keyword{aermod}
\keyword{control}
\keyword{input}
\keyword{options}

