% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downwash_tbl.R
\name{downwash_tbl}
\alias{downwash_tbl}
\title{Downwash options}
\usage{
downwash_tbl(STACK = "STK_1", BUILDHGTS = "", BUILDWIDS = "",
  BUILDLENS = "", XBADJ = "", YBADJ = "")
}
\arguments{
\item{STACK}{Stack names. Separate multiple sources with commas.}
}
\description{
Create an input table of AERMOD downwash options.
}
\examples{
downwash_tbl(STACK = "STK_1")
}
\keyword{aermod}
\keyword{building}
\keyword{downwash}
\keyword{input}

