\name{grid.polar}
\alias{grid.polar}
\title{grid.polar}
\usage{grid.polar(r, n, offset, increment, default.units="degrees", origin=c(0, 0))
grid.polar(r, n, offset, increment, default.units="degrees", origin=c(0, 0))
grid.polar(r, n, offset, increment, default.units="degrees", origin=c(0, 0))
}
\description{Generate a polar grid.}
\value{a SpatialPointsDataFrame\code{grid.polar}: a SpatialPointsDataFrame


\code{grid.polar}: a SpatialPointsDataFrame


}
\references{EPA (2004). User's Guide for the AMS/EPA Regulatory Model AERMOD. EPA-454/B-03-001.EPA (2004). User's Guide for the AMS/EPA Regulatory Model AERMOD. EPA-454/B-03-001.EPA (2004). User's Guide for the AMS/EPA Regulatory Model AERMOD. EPA-454/B-03-001.}
\arguments{\item{r}{distances (arbitrary units)}
\item{n}{angles}
\item{offset}{starting angle}
\item{increment}{angular increment}
\item{default.units}{if offset or increment do not have units, assume these}
\item{origin}{(optional) x, y offset}
}

\alias{grid.polar}

\alias{grid.polar}

