% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_tbl.R
\name{met_tbl}
\alias{met_tbl}
\title{Meteorology options}
\usage{
met_tbl(SURFFILE = "..\\\\MET\\\\RSTMPX5Y.SFC",
  PROFFILE = "..\\\\MET\\\\RSTMPX5Y.PFL",
  SURFDATA = "14925 2006 ROCHESTER/MUNICIPAL_ARPT", UAIRDATA = "94983 2006",
  PROFBASE = "396.0 METERS", STARTEND = "2009 1 1 1, 2013 12 31 24")
}
\arguments{
\item{SURFFILE}{Surface met data file location.}

\item{PROFFILE}{Profile met data file location.}
}
\description{
Create an input table of AERMOD meteorology options.
}
\examples{
met_tbl(SURFFILE = "..\\\\MET\\\\RSTMPX5Y.SFC")
}
\keyword{aermod}
\keyword{input}
\keyword{meteorology}
\keyword{surface}
\keyword{weather}

