% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_aermod.R
\name{new_aermod}
\alias{new_aermod}
\title{Create AERMOD input tables}
\usage{
new_aermod(control = "control", sources = "sources", meteorology = "met",
  output = "out")
}
\arguments{
\item{control}{Name for control options data frame. Default is "control".}

\item{sources}{Name for emission source data frame. Default is "sources".}

\item{meteorology}{Name for meteorology options data frame. Default is "met".}

\item{output}{Name for output options data frame. Default is "output".}
}
\description{
Create default AERMOD input data frames: 
(1) control 
(2) sources
(3) meteorology
(4) output_options
}
\examples{


new_aermod()
}
\keyword{aermod}
\keyword{input}
\keyword{new}
\keyword{start}
\keyword{tables}

