% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_tbl.R
\name{out_tbl}
\alias{out_tbl}
\title{Output options}
\usage{
out_tbl(RECTABLE = "ALLAVE 1ST, 1 1ST", MAXTABLE = "", DAYTABLE = "",
  PLOTFILE = "")
}
\arguments{
\item{RECTABLE}{Output options as text. Multiple options separated by a comma:}
}
\description{
Create an input table of AERMOD output options.
}
\examples{
out_tbl(RECTABLE = "ALLAVE 1ST, 1 1ST 8TH")
}
\keyword{aermod}
\keyword{input}
\keyword{output}

