% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_tbl.R
\name{out_tbl}
\alias{out_tbl}
\title{Output options}
\usage{
out_tbl(rect_table = c("ALLAVE 1ST", "1 1ST"), max_table = as.character(NA),
  day_table = as.character(NA), file_form = as.character(NA),
  rank_file = as.character(NA), plot_file = as.character(NA))
}
\arguments{
\item{rect_table}{Output options as text. Enter multiple options as a vector: \code{c("ALLAVE 1ST", "1 1ST")}}

\item{max_table}{Output options as text.}

\item{day_table}{Output options as text.}

\item{rank_file}{Output options as text.}

\item{plot_file}{Output options as text.}

\item{file_from}{Output options as text.}
}
\description{
Create a data frame of AERMOD output options.
}
\examples{
out_tbl(rect_table = c("ALLAVE 1ST", "1 1ST 8TH"))
}
\keyword{aermod}
\keyword{output}

