% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_tbl.R
\name{proj_tbl}
\alias{proj_tbl}
\title{Projection options}
\usage{
proj_tbl(projection = "Coordinate System UTM",
  description = "UTM: Universal Transverse Mercator",
  datum = "North American datum 1983", dtmrgn = "conus", units = "m",
  zone = "15", zoneinx = "0")
}
\arguments{
\item{projection}{Coordinate projection for AERMOD run.}

\item{zone}{UTM zone of coordinates.}
}
\description{
Create an input table of AERMOD projection options.
}
\examples{
proj_tbl(projection  = "Coordinate System UTM")
}
\keyword{aermod}
\keyword{input}
\keyword{options}
\keyword{projection}

