% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aermod_inp.R
\name{read_aermod_inp}
\alias{read_aermod_inp}
\title{Read AERMOD input file}
\usage{
read_aermod_inp(file = "aermod.inp")
}
\arguments{
\item{file}{File location. Default is "aermod.inp" in the working directory.}
}
\description{
Read an aermod.inp file into an AERMOD input table.
}
\examples{
read_aermod_inp(file = "aermod.inp")
}
\keyword{aermod}
\keyword{input}
\keyword{read}

