% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_df.R
\name{receptor_df}
\alias{receptor_df}
\title{Receptor options data frame}
\usage{
receptor_df(
  receptor_file = as.character(NA),
  receptor_as_text = as.character(NA)
)
}
\arguments{
\item{receptor_file}{Receptor file location.}

\item{receptor_as_text}{Additional receptor locations as text string.}
}
\description{
Create a data frame of AERMOD receptor options.
}
\examples{
\dontrun{receptor_df(receptor_file = "../receptors/circle_receptors.ROU")}

}
\keyword{aermod}
\keyword{input}
\keyword{receptors}
