% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aermod.R
\name{run_aermod}
\alias{run_aermod}
\title{Run AERMOD}
\usage{
run_aermod(
  input = "aermod.inp",
  out_file = "aermod.out",
  exe_folder = "aermod",
  silent = FALSE
)
}
\arguments{
\item{input}{Path to input file.}

\item{out_file}{File name for saving model output. 
Defaults to value of \code{title} in the provided input file.}

\item{exe_folder}{Folder containing \code{aermod.exe}.}

\item{silent}{Print results to screen. \code{TRUE} or \code{FALSE}}
}
\description{
Call aermod.exe and run input file
}
\examples{
\dontrun{
run_aermod(data       = "aermod.inp", 
           out_file   = "aermod_results", 
           exe_folder = "aermod_exe")
}
}
\keyword{EPA}
\keyword{aermod}
\keyword{dispersion}
\keyword{model}
\keyword{run}
