% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_postfile.R
\name{scan_postfile}
\alias{scan_postfile}
\title{Scan POST file}
\usage{
scan_postfile(f, ...)
}
\arguments{
\item{f}{path to an unformatted Fortran I/O binary POST file generated by AERMOD.}

\item{...}{arguments passed thru to \code{readBin}. Primary use is to specify endianness in case POST file was generated on a machine with different endianness from your own.}
}
\value{
matrix
}
\description{
Reads Fortran unformatted I/O POSTFILE and conver to matrix.
Hours are not written individually, nor are source groups
as they are saved as attributes of the matrix.
}
