% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_tbl.R
\name{source_tbl}
\alias{source_tbl}
\title{Source options}
\usage{
source_tbl(id = c("SV01", "SV02"), description = c("boiler1", "boiler2"),
  elevation_m = c(361.33, 361.33), x_coord = c(461874.9, 471977.5),
  y_coord = c(4631367, 4811367), type = c("point", "point"),
  emissions = c(1, 1), emis_unit = "g/s", height_m = c(10, 12),
  temp_k = c(600, 610), velocity_ms = c(30, 25), diameter_m = c(2, 2),
  downwash_file = "", group_id = "")
}
\arguments{
\item{id}{Source ID. For multiple sources use a vector: \code{c("SV01", "SV02")}.}

\item{elevation_m}{Source elevation in meters. For multiple sources use a vector: \code{c(333, 334)}.}

\item{x_coord}{X coordinates of source. For multiple sources use a vector: \code{c(461874.9, 471977.5)}.}

\item{y_coord}{Y coordinates of source. For multiple sources use a vector: \code{c(5631367, 5811367)}.}

\item{type}{Type of source to be modeled as in AERMOD. 
Options: (1) "point", (2) "volume", or (3) "area". 
For multiple sources use a vector: \code{c("point", "volumne")}.}

\item{emissions}{Emission rate for each source. For multiple sources use a vector: \code{c(1, 2)}.}

\item{height_m}{Height of source in meters. For multiple sources use a vector: \code{c(10, 10)}}

\item{temp_k}{Exit temperature of air emissions in Kelvin. For multiple sources use a vector: \code{c(610, 600)}.}

\item{velocity_ms}{Exit velocity of air emissions in in m/s.
Options: (1) "g/s", (2) "lb/hr", or (3) "tons/yr". 
For multiple sources use a vector: \code{c("g/s", "lb/hr")}.}

\item{diameter_m}{Source diameter in meters. For multiple sources use a vector: \code{c(1, 1.1)}.}

\item{group_id}{Group ID assigned to source. 
To assign multiple groups use a vector: \code{c("All", "SV01", "SV01_to_SV03")}.
If blank, sources are automatically assigned to group "All", 
which will contain the aggregated AERMOD results for all sources.}

\item{emis_units}{Units of measurement for \code{emissions}.
Options: (1) "g/s", (2) "lb/hr", or (3) "tons/yr". 
For multiple sources use a vector: \code{c("g/s", "lb/hr")}.}
}
\description{
Create an input table of AERMOD source options.
}
\examples{
source_tbl(id = "Stack_1", elevation = 300)
}
\keyword{aermod}
\keyword{input}
\keyword{source}

